data-ingestion-service/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
aiokafka==0.8.0
kafka-python==2.0.2
pydantic==2.5.0
python-dotenv==1.0.0
prometheus-client==0.19.0
httpx==0.25.0
